# Conflicts:
#	docs/tasks.md
#	mteb/abstasks/AbsTaskSpeedTask.py
#	mteb/abstasks/TaskMetadata.py
#	mteb/encoder_interface.py
#	mteb/models/misc_models.py
#	mteb/models/ops_moa_models.py
#	mteb/models/ru_sentence_models.py
#	mteb/models/sentence_transformers_models.py
#	mteb/tasks/Classification/__init__.py
#	mteb/tasks/Clustering/deu/BlurbsClusteringP2P.py
#	mteb/tasks/Clustering/deu/BlurbsClusteringS2S.py
#	mteb/tasks/Clustering/deu/TenKGnadClusteringS2S.py
#	mteb/tasks/Clustering/fra/AlloProfClusteringP2P.py
#	mteb/tasks/Clustering/fra/AlloProfClusteringS2S.py
#	mteb/tasks/Clustering/fra/HALClusteringS2S.py
#	mteb/tasks/Image/Any2AnyMultiChoice/eng/BLINKIT2IMultiChoice.py
#	mteb/tasks/Image/Any2AnyMultiChoice/eng/BLINKIT2TMultiChoice.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/BLINKIT2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/BLINKIT2TRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/CUB200I2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/FORBI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/GLDv2I2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/GLDv2I2TRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/ImageCoDeT2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/METI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/ROxfordI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/RP2kI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/RParisI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/SOPI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/SketchyI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/StanfordCarsI2IRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/VQA2IT2TRetrieval.py
#	mteb/tasks/Image/Any2AnyRetrieval/eng/VizWizIT2TRetrieval.py
#	mteb/tasks/Image/Clustering/eng/__init__.py
#	mteb/tasks/Image/ImageClassification/eng/BirdsnapClassification.py
#	mteb/tasks/Image/ImageClassification/eng/CIFAR.py
#	mteb/tasks/Image/ImageClassification/eng/Caltech101Classification.py
#	mteb/tasks/Image/ImageClassification/eng/DTDClassification.py
#	mteb/tasks/Image/ImageClassification/eng/FER2013Classification.py
#	mteb/tasks/Image/ImageClassification/eng/FGVCAircraftClassification.py
#	mteb/tasks/Image/ImageClassification/eng/Food101Classification.py
#	mteb/tasks/Image/ImageClassification/eng/GTSRBClassification.py
#	mteb/tasks/Image/ImageClassification/eng/MNISTClassification.py
#	mteb/tasks/Image/ImageClassification/eng/OxfordPetsClassification.py
#	mteb/tasks/Image/ImageClassification/eng/RESISC45Classification.py
#	mteb/tasks/Image/ImageClassification/eng/STL10Classification.py
#	mteb/tasks/Image/ImageClassification/eng/SUN397Classification.py
#	mteb/tasks/Image/ImageClassification/eng/StanfordCarsClassification.py
#	mteb/tasks/Image/ImageClassification/eng/UCF101Classification.py
#	mteb/tasks/Image/ImageClustering/eng/CIFAR.py
#	mteb/tasks/Image/ImageClustering/eng/ImageNet.py
#	mteb/tasks/Image/ImageMultilabelClassification/eng/PascalVOC2007.py
#	mteb/tasks/Image/ImageTextPairClassification/ImageCoDe.py
#	mteb/tasks/Image/VisualSTS/__init__.py
#	mteb/tasks/Image/VisualSTS/en/__init__.py
#	mteb/tasks/Image/ZeroShotClassification/eng/Birdsnap.py
#	mteb/tasks/Image/ZeroShotClassification/eng/CIFAR.py
#	mteb/tasks/Image/ZeroShotClassification/eng/Caltech101.py
#	mteb/tasks/Image/ZeroShotClassification/eng/DTD.py
#	mteb/tasks/Image/ZeroShotClassification/eng/EuroSAT.py
#	mteb/tasks/Image/ZeroShotClassification/eng/FER2013.py
#	mteb/tasks/Image/ZeroShotClassification/eng/FGVCAircraft.py
#	mteb/tasks/Image/ZeroShotClassification/eng/Food101.py
#	mteb/tasks/Image/ZeroShotClassification/eng/GTSRB.py
#	mteb/tasks/Image/ZeroShotClassification/eng/MNIST.py
#	mteb/tasks/Image/ZeroShotClassification/eng/OxfordPets.py
#	mteb/tasks/Image/ZeroShotClassification/eng/PatchCamelyon.py
#	mteb/tasks/Image/ZeroShotClassification/eng/RESISC45.py
#	mteb/tasks/Image/ZeroShotClassification/eng/STL10.py
#	mteb/tasks/Image/ZeroShotClassification/eng/SUN397.py
#	mteb/tasks/Image/__init__.py
#	mteb/tasks/MultiLabelClassification/__init__.py
#	mteb/tasks/Reranking/zho/CMTEBReranking.py
#	mteb/tasks/Retrieval/__init__.py
#	mteb/tasks/__init__.py
#	pyproject.toml
#	tests/test_TaskMetadata.py

+++ mteb/models/jasper_models.py
@@ -66,7 +66,7 @@ jasper_en_v1 = ModelMeta(
         config_kwargs={"is_text_encoder": True, "vector_dim": 12288},
         model_kwargs={
             "attn_implementation": "sdpa",
-            "torch_dtype": torch.float16,
+            "torch_dtype": torch.bfloat16,
         },
         trust_remote_code=True,
         max_seq_length=2048,
